requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
alabaster==1.0.0
annotated-types==0.7.0
attrs==25.3.0
babel==2.17.0
basedpyright==1.30.1
blinker==1.8.2
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.1.7
docutils==0.21.2
flask==3.0.3
flask-paginate==2024.4.12
flask-wtf==1.2.2
frozenlist==1.7.0
greenlet==3.2.3
idna==3.10
imagesize==1.4.1
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.4
libsql-client==0.3.1
libsql-experimental==0.0.55
markupsafe==3.0.2
multidict==6.6.3
nodejs-wheel-binaries==22.17.0
packaging==25.0
propcache==0.3.2
pydantic==2.9.2
pydantic-core==2.23.4
pygments==2.19.2
python-decouple==3.8
requests==2.32.4
roman-numerals-py==3.1.0
ruff==0.12.2
snowballstemmer==3.0.1
sphinx==8.2.3
sphinx-press-theme==0.8.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlalchemy==2.0.41
sqlalchemy-libsql==0.2.0
typing-extensions==4.12.2
urllib3==2.5.0
werkzeug==3.1.2
wtforms==3.2.1
yarl==1.20.1
zipp==3.21.0
